pyproject.toml
[tool.poetry]
name = "exako"
version = "0.1.0"
description = ""
authors = ["Gabriel"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlmodel = "^0.0.22"
pydantic-settings = "^2.5.2"
fastapi-pagination = "^0.12.31"
fief-client = {extras = ["fastapi"], version = "^0.20.0"}
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.9"
isort = "^5.13.2"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
pybabel = "^0.0.0.dev0"
alembic = "^1.13.3"

[tool.ruff]
exclude = [".venv", "migrations"]

[tool.ruff.format]
quote-style = "single"

[tool.isort]
profile = "black"
extend_skip = ["migrations"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
format = "ruff format . && isort ."
run = "fastapi dev exako/main.py --port 8080"
pre_test = "task format"
test = "pytest -s --cov=. -vv"
post_test = "coverage html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
